+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - update UglifyJS to 2.8.0
+- enable collapse_vars compress options by default
 
 ## 3.0.4 (30 November 2016)